+++ doorstop/core/item.py
@@ -10,11 +10,10 @@ from typing import Any, List
 import pyficache
 
 from doorstop import common, settings
-from doorstop.common import DoorstopError, DoorstopInfo, DoorstopWarning
+from doorstop.common import DoorstopError
 from doorstop.core import editor
 from doorstop.core.base import (
     BaseFileObject,
-    BaseValidatable,
     add_item,
     auto_load,
     auto_save,
@@ -66,7 +65,7 @@ def requires_tree(func):
     return wrapped
 
 
-class Item(BaseValidatable, BaseFileObject):  # pylint: disable=R0902
+class Item(BaseFileObject):  # pylint: disable=R0902
     """Represents an item file with linkable text."""
 
     EXTENSIONS = '.yml', '.yaml'
@@ -553,184 +552,11 @@ class Item(BaseValidatable, BaseFileObject):  # pylint: disable=R0902
         except KeyError:
             log.warning("link to {0} does not exist".format(uid))
 
-    def get_issues(
-        self, skip=None, document_hook=None, item_hook=None
-    ):  # pylint: disable=unused-argument
-        """Yield all the item's issues.
+    def is_reviewed(self):
+        return not self._data['reviewed']
 
-        :param skip: list of document prefixes to skip
-
-        :return: generator of :class:`~doorstop.common.DoorstopError`,
-                              :class:`~doorstop.common.DoorstopWarning`,
-                              :class:`~doorstop.common.DoorstopInfo`
-
-        """
-        assert document_hook is None
-        assert item_hook is None
-        skip = [] if skip is None else skip
-
-        log.info("checking item %s...", self)
-
-        # Verify the file can be parsed
-        self.load()
-
-        # Skip inactive items
-        if not self.active:
-            log.info("skipped inactive item: %s", self)
-            return
-
-        # Delay item save if reformatting
-        if settings.REFORMAT:
-            self.auto = False
-
-        # Check text
-        if not self.text:
-            yield DoorstopWarning("no text")
-
-        # Check external references
-        if settings.CHECK_REF:
-            try:
-                self.find_ref()
-            except DoorstopError as exc:
-                yield exc
-
-        # Check links
-        if not self.normative and self.links:
-            yield DoorstopWarning("non-normative, but has links")
-
-        # Check links against the document
-        yield from self._get_issues_document(self.document, skip)
-
-        if self.tree:
-            # Check links against the tree
-            yield from self._get_issues_tree(self.tree)
-
-            # Check links against both document and tree
-            yield from self._get_issues_both(self.document, self.tree, skip)
-
-        # Check review status
-        if not self.reviewed:
-            if settings.CHECK_REVIEW_STATUS:
-                if not self._data['reviewed']:
-                    if settings.REVIEW_NEW_ITEMS:
-                        self.review()
-                    else:
-                        yield DoorstopInfo("needs initial review")
-                else:
-                    yield DoorstopWarning("unreviewed changes")
-
-        # Reformat the file
-        if settings.REFORMAT:
-            log.debug("reformatting item %s...", self)
-            self.save()
-
-    def _get_issues_document(self, document, skip):
-        """Yield all the item's issues against its document."""
-        log.debug("getting issues against document...")
-
-        if document in skip:
-            log.debug("skipping issues against document %s...", document)
-            return
-
-        # Verify an item's UID matches its document's prefix
-        if self.prefix != document.prefix:
-            msg = "prefix differs from document ({})".format(document.prefix)
-            yield DoorstopInfo(msg)
-
-        # Verify that normative, non-derived items in a child document have at
-        # least one link.  It is recommended that these items have an upward
-        # link to an item in the parent document, however, this is not
-        # enforced.  An info message is generated if this is not the case.
-        if all((document.parent, self.normative, not self.derived)) and not self.links:
-            msg = "no links to parent document: {}".format(document.parent)
-            yield DoorstopWarning(msg)
-
-        # Verify an item's links are to the correct parent
-        for uid in self.links:
-            try:
-                prefix = uid.prefix
-            except DoorstopError:
-                msg = "invalid UID in links: {}".format(uid)
-                yield DoorstopError(msg)
-            else:
-                if document.parent and prefix != document.parent:
-                    # this is only 'info' because a document is allowed
-                    # to contain items with a different prefix, but
-                    # Doorstop will not create items like this
-                    msg = "parent is '{}', but linked to: {}".format(
-                        document.parent, uid
-                    )
-                    yield DoorstopInfo(msg)
-
-    def _get_issues_tree(self, tree):
-        """Yield all the item's issues against its tree."""
-        log.debug("getting issues against tree...")
-
-        # Verify an item's links are valid
-        identifiers = set()
-        for uid in self.links:
-            try:
-                item = tree.find_item(uid)
-            except DoorstopError:
-                identifiers.add(uid)  # keep the invalid UID
-                msg = "linked to unknown item: {}".format(uid)
-                yield DoorstopError(msg)
-            else:
-                # check the linked item
-                if not item.active:
-                    msg = "linked to inactive item: {}".format(item)
-                    yield DoorstopInfo(msg)
-                if not item.normative:
-                    msg = "linked to non-normative item: {}".format(item)
-                    yield DoorstopWarning(msg)
-                # check the link status
-                if uid.stamp == Stamp(True):
-                    uid.stamp = item.stamp()
-                elif not str(uid.stamp) and settings.STAMP_NEW_LINKS:
-                    uid.stamp = item.stamp()
-                elif uid.stamp != item.stamp():
-                    if settings.CHECK_SUSPECT_LINKS:
-                        msg = "suspect link: {}".format(item)
-                        yield DoorstopWarning(msg)
-                # reformat the item's UID
-                identifier2 = UID(item.uid, stamp=uid.stamp)
-                identifiers.add(identifier2)
-
-        # Apply the reformatted item UIDs
-        if settings.REFORMAT:
-            self._data['links'] = identifiers
-
-    def _get_issues_both(self, document, tree, skip):
-        """Yield all the item's issues against its document and tree."""
-        log.debug("getting issues against document and tree...")
-
-        if document.prefix in skip:
-            log.debug("skipping issues against document %s...", document)
-            return
-
-        # Verify an item is being linked to (child links)
-        if settings.CHECK_CHILD_LINKS and self.normative:
-            find_all = settings.CHECK_CHILD_LINKS_STRICT or False
-            items, documents = self._find_child_objects(
-                document=document, tree=tree, find_all=find_all
-            )
-
-            if not items:
-                for child_document in documents:
-                    if document.prefix in skip:
-                        msg = "skipping issues against document %s..."
-                        log.debug(msg, child_document)
-                        continue
-                    msg = "no links from child document: {}".format(child_document)
-                    yield DoorstopWarning(msg)
-            elif settings.CHECK_CHILD_LINKS_STRICT:
-                prefix = [item.prefix for item in items]
-                for child in document.children:
-                    if child in skip:
-                        continue
-                    if child not in prefix:
-                        msg = 'no links from document: {}'.format(child)
-                        yield DoorstopWarning(msg)
+    def set_links(self, links):
+        self._data['links'] = links
 
     @requires_tree
     def find_ref(self):
@@ -788,7 +614,7 @@ class Item(BaseValidatable, BaseFileObject):  # pylint: disable=R0902
         :return: list of found item UIDs
 
         """
-        items, _ = self._find_child_objects(find_all=find_all)
+        items, _ = self.find_child_items_and_documents(find_all=find_all)
         identifiers = [item.uid for item in items]
         return identifiers
 
@@ -802,7 +628,7 @@ class Item(BaseValidatable, BaseFileObject):  # pylint: disable=R0902
         :return: list of found items
 
         """
-        items, _ = self._find_child_objects(find_all=find_all)
+        items, _ = self.find_child_items_and_documents(find_all=find_all)
         return items
 
     child_items = property(find_child_items)
@@ -813,12 +639,12 @@ class Item(BaseValidatable, BaseFileObject):  # pylint: disable=R0902
         :return: list of found documents
 
         """
-        _, documents = self._find_child_objects(find_all=False)
+        _, documents = self.find_child_items_and_documents(find_all=False)
         return documents
 
     child_documents = property(find_child_documents)
 
-    def _find_child_objects(self, document=None, tree=None, find_all=True):
+    def find_child_items_and_documents(self, document=None, tree=None, find_all=True):
         """Get lists of child items and child documents.
 
         :param document: document containing the current item
